fix: 잔액 타입 변경 / 송금

+++ backend/src/main/kotlin/com/finance/backend/bank/response/BankAccountRes.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 
 class BankAccountRes(
         acNo : String,
-        balance : Int,
+        balance : Long,
         acName : String,
         cpName: String,
         cpLogo: String
@@ -14,7 +14,7 @@ class BankAccountRes(
         @JsonProperty("ac_no")
         val acNo: String = acNo
 
-        val balance: Int = balance
+        val balance: Long = balance
 
         @JsonProperty("ac_name")
         val acName: String = acName
